+++ tests/testthat/testISRates.R
@@ -8,9 +8,10 @@ test_that("isrates and CIs calculate correctly",{
 
   expect_equal(data.frame(select(calculate_ISRate(select(test_ISR_ownref,-refcount,-refpop), total_count, pop,
                                                   x_ref = test_ISR_refdata$refcount, n_ref = test_ISR_refdata$refpop,
-                                                  x_lookup = test_ISR_lookup, lookup_cols = "area"), 1:7,9:10)),
+                                                  observed_totals = test_ISR_lookup, #lookup_cols = "area"
+                                                  ), 1:7,9:10)),
                data.frame(select(slice(test_ISR_results,1:3),1:7,10:11)),
-               check.attributes=FALSE, check.names=FALSE,info="test default with x_lookup")
+               check.attributes=FALSE, check.names=FALSE,info="test default with observed_totals")
 
   expect_equal(data.frame(select(calculate_ISRate(select(test_ISR_ownref,-refcount,-refpop), count, pop, confidence = c(0.95,0.998),
                                          x_ref = test_ISR_refdata$refcount, n_ref = test_ISR_refdata$refpop),1:9,11:12)),
@@ -167,12 +168,12 @@ test_that("isrates - errors are generated when invalid arguments are used",{
                info="error invalid number of arguments")
 
   expect_error(calculate_ISRate(test_ISR_ownref, total_count, pop, x_ref = test_ISR_refdata$refcount,
-                                n_ref = test_ISR_refdata$refpop, x_lookup = test_ISR_lookup, lookup_cols="areas"),
+                                n_ref = test_ISR_refdata$refpop, observed_totals = test_ISR_lookup, lookup_cols="areas"),
                "lookup_cols are not in data",info="wrong column names for data")
 
   expect_error(calculate_ISRate(test_ISR_ownref, total_count, pop, x_ref = test_ISR_refdata$refcount,
-                                n_ref = test_ISR_refdata$refpop, x_lookup = test_ISR_lookup, lookup_cols="ageband"),
-               "lookup_cols are not in x_lookup",info="wrong column names for x lookup")
+                                n_ref = test_ISR_refdata$refpop, observed_totals = test_ISR_lookup, lookup_cols="ageband"),
+               "lookup_cols are not in observed_totals",info="wrong column names for x lookup")
 
 
 })